+++ src/events/clientReady.ts
@@ -3,8 +3,8 @@ import { exit } from 'node:process';
 import { Events } from 'discord.js';
 
 import { EventHandler } from './EventHandler.ts';
-import { syncCommands } from '../commands/index.ts';
 import { debug, error, info } from '../logger.ts';
+import { syncCommands } from '../syncCommands.ts';
 
 /** Called once the client successfully logs in. */
 export const clientReady = new EventHandler(Events.ClientReady)
